--- a/apps/routes/cutouts/utils.py
+++ b/apps/routes/cutouts/utils.py
@@ -91,7 +91,7 @@ def format_and_send_cutout(payload: dict):
         group_alerts=False,
         truncated=True,
         extract_color=False,
-        escape_slash=True,
+        escape_slash=False,
     )
 
     json_payload = {}